+++ kernel/src/driver/disk/ahci/sata_disk.h
@@ -14,7 +14,7 @@ bool sata_read(HBA_PORT_T* port, size_t _lba, size_t _count, uintptr_t _buf_phys
 bool sata_write(HBA_PORT_T* port, size_t _lba, size_t _count, uintptr_t _phys_buf_addr);
 
 uint64_t sata_get_total_sectors(HBA_PORT_T* port);
-uint32_t sata_get_bytes_per_sector(HBA_PORT_T* port);
+uint16_t sata_get_bytes_per_sector(HBA_PORT_T* port);
 
 void sata_disk_identify(HBA_PORT_T* port);
